requirements.txt
Pillow==8.1.2
django-mathfilters==1.0.0